test: separate lazy offload manager contracts
